requirements.txt
Pillow==9.5.0
Pillow==10.0.0
pytesseract==0.3.10
